+++ endcheck.h
@@ -3,7 +3,7 @@
 #ifndef ENDCHECK_H
 #define ENDCHECK_H
 
-void checkEnd(short happiness, int techDiscovered, unsigned long population, unsigned long landArea);
+void checkEnd(short happiness, short techDiscovered, unsigned long population, unsigned long landArea);
 void lose(unsigned short ending);
 void win (unsigned short ending);
