+++ turn.h
@@ -6,7 +6,7 @@
 #ifndef TURN_H
 #define TURN_H
 
-std::vector<int> nextTurn(unsigned long population, unsigned long landArea, int techDiscovered, short happiness, std::string name, bool defaultWork);
-std::vector<double> findMultiplier(int techDiscovered);
+std::vector<long long int> nextTurn(unsigned long population, unsigned long landArea, short techDiscovered, short happiness, std::string name, bool defaultWork);
+std::vector<double> findMultiplier(short techDiscovered);
 
 #endif
